+++ CHANGELOG.md
@@ -76,6 +76,7 @@ disables commands being added to history.
 random errors introduced with the new Lua runtime (see [#197])
 - `bait.release(name, catcher)` removes `handler` for the named `event`
 - `exec`, `clear` and `cat` builtin commands
+- `hilbish.cancel` hook
 
 [#197]: https://github.com/Rosettea/Hilbish/issues/197